Handle BCR data import, update sample data
