+++ test/sampledata/dataspace/bcrimport/allele_sequence.csv
@@ -2,8 +2,8 @@ allele,allele_sequence_nt,release_date
 GGGG7-16*01,TTTTAAAGTCGGGGGCTGCCGTGAGGAAGCCTGGGGCCCT,1/20/2023
 VVVV7-51*07,GCCCGGGGCATGGATGAACCCTACAAATGGAG,1/20/2023
 RRRR31-37*01,CCCCCTTTTTTTTTTTTTTACAGGAGTCGGGCCTTCCCCGGACTGGTGAAGCCTTCGGAAACCCTGTCCCTCACGTGCAGTGTTTT,1/20/2023
-POPOK32*02,AACGGTCCCCAGGGAAGGGTTTACAGG,1/20/2023
-RRRR31-37*01,CGGGAAACCCCGGACTGGTGAAGCCTTCGGGGG,1/20/2023
+POPO32*02,AACGGTCCCCAGGGAAGGGTTTACAGG,1/20/2023
+RRRS31A-37*01,CGGGAAACCCCGGACTGGTGAAGCCTTCGGGGG,1/20/2023
 RRRR31D-37*01,GGGCAGCCTCCACGACTCCTAAGGGGTGTATTTGGCC,1/20/2023
 POPOK32*01,CAAGAAGTCAGAGCCTCCTGTGGAATAAGAACGAT,1/20/2023
 POPOK32*02,CAGAAAGTCAGAGCCTCCTGTGGAATAAGGGCAGT,1/20/2023
