+++ test/sampledata/dataspace/bcrimport/run_log.csv
@@ -1,0 +1,3 @@
+run_number,run_application,run_information
+20230202_igblast,,
+20230202_vquest,,
